+++ ktlint-test/src/main/kotlin/com/pinterest/ktlint/test/RuleExtension.kt
@@ -1,13 +1,39 @@
 package com.pinterest.ktlint.test
 
+import com.pinterest.ktlint.core.KTLINT_UNIT_TEST_DUMP_AST
+import com.pinterest.ktlint.core.KTLINT_UNIT_TEST_ON_PROPERTY
 import com.pinterest.ktlint.core.KtLint
 import com.pinterest.ktlint.core.LintError
 import com.pinterest.ktlint.core.Rule
 import com.pinterest.ktlint.core.RuleSet
-import java.util.ArrayList
+import com.pinterest.ktlint.core.initKtLintKLogger
+import com.pinterest.ruleset.test.DumpASTRule
+import mu.KotlinLogging
 import org.assertj.core.api.Assertions
 import org.assertj.core.util.diff.DiffUtils.diff
 import org.assertj.core.util.diff.DiffUtils.generateUnifiedDiff
+import org.jetbrains.kotlin.utils.addToStdlib.ifTrue
+
+private val logger = KotlinLogging.logger {}.initKtLintKLogger()
+
+private fun Rule.toRuleSets(): List<RuleSet> {
+    val dumpAstRuleSet = System
+        .getenv(KTLINT_UNIT_TEST_DUMP_AST)
+        .orEmpty()
+        .equals(KTLINT_UNIT_TEST_ON_PROPERTY, ignoreCase = true)
+        .ifTrue {
+            logger.info { "Dump AST of code before processing as System environment variable $KTLINT_UNIT_TEST_DUMP_AST is set to 'on'" }
+            RuleSet(
+                "debug",
+                DumpASTRule(
+                    // Write to STDOUT. The focus in a failed unit test should first go to the error in the rule that is
+                    // to be tested and not to the AST,
+                    out = System.out
+                )
+            )
+        }
+    return listOfNotNull(dumpAstRuleSet, RuleSet("standard", this))
+}
 
 public fun Rule.lint(
     text: String,
@@ -22,21 +48,14 @@ public fun Rule.lint(
     script: Boolean = false
 ): List<LintError> {
     val res = ArrayList<LintError>()
-    val debug = debugAST()
     KtLint.lint(
         KtLint.Params(
             fileName = lintedFilePath,
             text = text,
-            ruleSets = (if (debug) listOf(RuleSet("debug", DumpAST())) else emptyList()) +
-                listOf(RuleSet("standard", this@lint)),
+            ruleSets = this.toRuleSets(),
             userData = userData,
             script = script,
-            cb = { e, _ ->
-                if (debug) {
-                    System.err.println("^^ lint error")
-                }
-                res.add(e)
-            }
+            cb = { e, _ -> res.add(e) }
         )
     )
     return res
@@ -59,8 +78,7 @@ public fun Rule.format(
     KtLint.Params(
         fileName = lintedFilePath,
         text = text,
-        ruleSets = (if (debugAST()) listOf(RuleSet("debug", DumpAST())) else emptyList()) +
-            listOf(RuleSet("standard", this@format)),
+        ruleSets = this.toRuleSets(),
         userData = userData,
         script = script,
         cb = cb