+++ ktlint/src/main/kotlin/com/pinterest/ktlint/internal/RuleSetsLoader.kt
@@ -1,8 +1,12 @@
 package com.pinterest.ktlint.internal
 
 import com.pinterest.ktlint.core.RuleSetProvider
+import com.pinterest.ktlint.core.initKtLintKLogger
 import java.net.URLClassLoader
 import java.util.ServiceLoader
+import mu.KotlinLogging
+
+private val logger = KotlinLogging.logger {}.initKtLintKLogger()
 
 /**
  * Load given list of paths to rulesets jars into map of ruleset providers.
@@ -27,10 +31,10 @@ internal fun JarFiles.loadRulesets(
     .filterKeys { !(disabledRules.isStandardRuleSetDisabled() && it == "\u0000standard") }
     .toSortedMap()
     .also { ruleSetMap ->
-        if (debug) {
-            ruleSetMap.forEach { entry ->
-                println("[DEBUG] Discovered ruleset with \"${entry.key}\" id.")
-            }
+        logger.debug {
+            ruleSetMap
+                .keys
+                .joinToString(prefix = "Discovered ruleset(s):\n\t", separator = "\n\t")
         }
 
         val expectedNumberOfCustomRuleSetsToBeLoaded = this.distinct().count()
@@ -41,18 +45,18 @@ internal fun JarFiles.loadRulesets(
                 .map { it.javaClass.canonicalName }
         val actualNumberOfCustomRuleSetsLoaded = customRuleSetsLoaded.count()
         if (expectedNumberOfCustomRuleSetsToBeLoaded != actualNumberOfCustomRuleSetsLoaded) {
-            System.err.println(
+            logger.warn {
                 """
-                [WARNING] Number of custom rule sets loaded does not match the expected number of rule sets to be loaded.
-                          Expected to load $expectedNumberOfCustomRuleSetsToBeLoaded custom rule set(s) for rules: ${this.joinToString()}
-                          Actually loaded $actualNumberOfCustomRuleSetsLoaded custom rule set(s) with names: ${customRuleSetsLoaded.joinToString() }
-                          One or more of the specified jars does not provide the custom rule set correctly. Check following:
-                            - Does the jar contain an implementation of the RuleSetProvider interface?
-                            - Does the jar contain a resource file with name "com.pinterest.ktlint.core.RuleSetProvider"?
-                            - Is the resource file located in directory "src/main/resources/META-INF/services"?
-                            - Does the resource file contain the fully qualified class name of the class implementing the RuleSetProvider interface?
+                Number of custom rule sets loaded does not match the expected number of rule sets to be loaded.
+                Expected to load $expectedNumberOfCustomRuleSetsToBeLoaded custom rule set(s) for rules: ${this.joinToString()}
+                Actually loaded $actualNumberOfCustomRuleSetsLoaded custom rule set(s) with names: ${customRuleSetsLoaded.joinToString()}
+                One or more of the specified jars does not provide the custom rule set correctly. Check following:
+                  - Does the jar contain an implementation of the RuleSetProvider interface?
+                  - Does the jar contain a resource file with name "com.pinterest.ktlint.core.RuleSetProvider"?
+                  - Is the resource file located in directory "src/main/resources/META-INF/services"?
+                  - Does the resource file contain the fully qualified class name of the class implementing the RuleSetProvider interface?
                 """.trimIndent() // ktlint-disable string-template
-            )
+            }
         }
     }